Overhaul of value fetching logic. Further view division.

+++ app/src/main/java/pl/cendrzak/szymon/SubaruSsm/SubaruValue.java
@@ -1,5 +1,5 @@
 package pl.cendrzak.szymon.SubaruSsm;
 
 public class SubaruValue {
-    public int value;
+    public Double value;
 }
